requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.9
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.8.2
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
cryptography==44.0.2
dacite==1.9.2
defusedxml==0.7.1
dill==0.4.0
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.5
django-cache-url==3.4.5
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-cron==0.6.0
django-crontab==0.7.1
django-extensions==4.1
django-filter==25.1
django-ipware==7.0.1
django-on-heroku==1.1.2
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
environs==14.1.1
factory_boy==3.3.3
Faker==37.1.0
fontawesomefree==6.0.0
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib==1.0.4
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.2
loguru==0.7.3
lxml==5.4.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
nose==1.3.7
oauthlib==3.2.2
packaging==25.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==3.1.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.2
psycopg==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pylint==3.3.6
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-ipware==3.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
python_unittest==0.2.15
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
ruff==0.11.7
sentry-sdk==2.26.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
stripe==12.1.0
testsolar-testtool-sdk==0.4.5
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unittest-xml-reporting==3.2.0
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
